tracing

+++ server/src/responses/mod.rs
@@ -5,6 +5,6 @@ pub mod auth;
 #[derive(Serialize)]
 pub struct PaginationResponse<T> {
     pub nodes: Vec<T>,
-    pub page: u16,
+    pub page: u64,
     pub total: u64,
 }
